+++ src/Resources/contao/config/config.php
@@ -1,15 +1,17 @@
 <?php
 
-/**
- * Hooks
+/*
+ * Copyright (c) 2023 Heimrich & Hannot GmbH
+ *
+ * @license LGPL-3.0-or-later
  */
-$GLOBALS['TL_HOOKS']['getPageLayout']['huh_amp'] = ['huh.amp.event_listener.hook_listener', 'getPageLayout'];
+
 $GLOBALS['TL_HOOKS']['parseTemplate']['huh_amp'] = [\HeimrichHannot\AmpBundle\EventListener\ParseTemplateListener::class, 'onParseTemplate'];
 $GLOBALS['TL_HOOKS']['generatePage']['huh_amp'] = [\HeimrichHannot\AmpBundle\EventListener\GeneratePageListener::class, 'onGeneratePage'];
 $GLOBALS['TL_HOOKS']['parseFrontendTemplate']['huh_amp'] = [\HeimrichHannot\AmpBundle\EventListener\ParseFrontendTemplateListener::class, 'onParseFrontendTemplate'];
 $GLOBALS['TL_HOOKS']['modifyFrontendPage']['huh_amp'] = ['huh.amp.event_listener.hook_listener', 'modifyFrontendPage'];
 
-/**
+/*
  * Frontend modules
  */
 System::getContainer()->get('huh.utils.array')->insertInArrayByName(